FIXED database on app products
